# openapi/ancf.v1.yaml
openapi: "3.0.3"
info:
  title: ANCF Commerce API v1
  description: |
    Agent-Native Commerce Framework (ANCF) v1 API.

    ANCF is a Zero-Frontend Commerce protocol where Agents discover shops,
    search products, obtain backend-authoritative quotes, and submit signed
    checkout intents. The backend is authoritative for all trusted state:
    pricing, inventory, ledger, minting, and provisioning.

    ## Agent Requirements

    Agents must:
    - Fetch and validate `/.well-known/agent-rules.json` before any commerce action.
    - Verify manifest signature, schema, and firmware SRI.
    - Call `/quote` after `/search`; never use search prices for checkout.
    - Send `Idempotency-Key` for all mutating APIs.
    - Require human confirmation before real asset movement.
    - Never call mint/redemption APIs unless the user explicitly chooses them.

    ## Security

    - All mutating APIs require an `Idempotency-Key` header.
    - `checkout/commit` requires a valid wallet signature over the canonical signable payload.
    - `X-API-Key` header is required for API authentication.
    - The backend verifies: quote freshness, idempotency, wallet signature,
      inventory availability, vUSDC balance, and Agent session policy.
  version: "1.0.0"
  contact:
    name: ANCF Team
    url: https://github.com/ancf-commerce/ancf
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.yourshop.com
...
paths:
...
  /api/v1/cli/search:
    get:
      tags: [Catalog]
      summary: Search products
      description: |
        Search for products by keyword. Results are display-only and must not be
        used directly for checkout pricing. Agents must call /api/v1/cli/quote
        to obtain an authoritative price before checkout.
      operationId: searchProducts
      parameters:
        - name: q
          in: query
          required: true
          description: Search query string
          schema:
            type: string
            minLength: 1
        - name: limit
          in: query
          required: false
          description: Maximum number of results to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          required: false
          description: Pagination offset
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
              example:
                items:
                  - sku_id: "sku_h100_v1"
                    title: "H100 compute rental, hourly"
                    price:
                      currency: "vUSDC"
                      amount_minor: "2450000"
                      scale: 6
                    stock_hint: 42
                    specs:
                      GPU: "80GB SXM5"
                      CUDA: "12.4"
                    media:
                      thumbnail: "https://cdn.yourshop.com/h100.png"
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
  /api/v1/cli/quote:
    post:
      tags: [Quote]
      summary: Get an authoritative price quote
      description: |
        Submit line items and receive a backend-authoritative quote with
        a short expiration window (default 5 minutes). Quotes are single-use;
        once consumed by a successful checkout commit, they cannot be reused.
      operationId: createQuote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuoteRequest'
            example:
              wallet: "USER_WALLET"
              network: "solana-mainnet"
              lines:
                - sku_id: "sku_h100_v1"
                  quantity: 2
      responses:
        "200":
          description: Authoritative quote
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteResponse'
              example:
                quote_id: "quote_01J..."
                currency: "vUSDC"
                total_minor: "4900000"
                scale: 6
                expires_at: "2026-06-04T00:10:00Z"
                lines:
                  - sku_id: "sku_h100_v1"
                    quantity: 2
                    unit_price_minor: "2450000"
                    line_total_minor: "4900000"
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authenticating requests. Required for all endpoints except /health.
  schemas:
...
    SearchResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SearchItem'
        total:
          type: integer
          minimum: 0
        limit:
          type: integer
          minimum: 1
          maximum: 100
        offset:
          type: integer
          minimum: 0
    SearchItem:
      type: object
      required: [sku_id, title, price]
      properties:
        sku_id:
          type: string
          minLength: 1
        title:
          type: string
          minLength: 1
        price:
          $ref: '#/components/schemas/DisplayPrice'
        stock_hint:
          type: integer
          minimum: 0
        specs:
          type: object
          additionalProperties:
            type: string
        media:
          $ref: '#/components/schemas/SearchMedia'
    DisplayPrice:
      type: object
      required: [currency, amount_minor, scale]
      properties:
        currency:
          type: string
          enum: [vUSDC]
        amount_minor:
          type: string
          pattern: '^[0-9]+$'
        scale:
          type: integer
          minimum: 0
    SearchMedia:
      type: object
      properties:
        thumbnail:
          type: string
          format: uri
        gallery:
          type: array
          items:
            type: string
            format: uri
    QuoteRequest:
      type: object
      required: [wallet, network, lines]
      properties:
        wallet:
          type: string
          minLength: 1
        network:
          type: string
          enum: [solana-mainnet, sonic-l2, solana-devnet]
        payment_rail:
          type: string
        lines:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/QuoteLineItem'
    QuoteLineItem:
      type: object
      required: [sku_id, quantity]
      properties:
        sku_id:
          type: string
          minLength: 1
        quantity:
          type: integer
          exclusiveMinimum: 0
    QuoteResponse:
      type: object
      required: [quote_id, currency, total_minor, scale, expires_at, lines]
      properties:
        quote_id:
          type: string
          pattern: '^quote_'
        currency:
          type: string
          enum: [vUSDC, CNY]
        total_minor:
          type: string
          pattern: '^[0-9]+$'
        scale:
          type: integer
          minimum: 0
        expires_at:
          type: string
          format: date-time
        lines:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/QuoteResponseLine'
    QuoteResponseLine:
      type: object
      required: [sku_id, quantity, unit_price_minor, line_total_minor]
      properties:
        sku_id:
          type: string
          minLength: 1
        quantity:
          type: integer
          exclusiveMinimum: 0
        unit_price_minor:
          type: string
          pattern: '^[0-9]+$'
        line_total_minor:
          type: string
          pattern: '^[0-9]+$'
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  reason:
                    type: string
            request_id:
              type: string
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "INVALID_PARAMETER"
              message: "Validation failed for field 'wallet'"
              request_id: "req_01J..."
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "UNAUTHORIZED"
              message: "Missing or invalid X-API-Key header"
              request_id: "req_01J..."
security:
  - ApiKeyAuth: []